+++ src/fields/AddressField.php
@@ -683,6 +683,11 @@ class AddressField extends Field implements PreviewableFieldInterface
      */
     public function getSearchKeywords(mixed $value, ElementInterface $element): string
     {
+        // If no value, bail with empty string
+        if (!$value) {
+            return '';
+        }
+
         /** @var AddressModel $value */
         return implode(' ', [
             $value->formatted,