+++ CHANGELOG.md
@@ -14,6 +14,11 @@
 ### Changed
 - Craft 5 compatibility.
 
+## Unreleased
+
+### Fixed
+- Added a null check when compiling search keywords. ([#108](https://github.com/doublesecretagency/craft-googlemaps/issues/108))
+
 ## 4.5.0 - 2024-03-21
 
 ### Added